split rotate function into new pointing_direction_rotation module, #589
