+++ tests/unit/measurement_set/_utils/_asdm/test_create_pointing_xds.py
@@ -1,7 +1,3 @@
-from contextlib import nullcontext as no_raises
-
-import numpy as np
-
 import pytest
 
 import pyasdm
@@ -10,7 +6,6 @@ import xarray as xr
 
 from xradio.measurement_set._utils._asdm.create_pointing_xds import (
     create_pointing_xds,
-    _rotate_offset_to_target,
 )
 from xradio.measurement_set.schema import PointingXds
 from xradio.schema.check import check_dataset
@@ -115,47 +110,3 @@ def test_create_pointing_xds_with_asdm_antenna_pointing(
     # The set of variables expected in this example
     for data_var in ["DIRECTION", "POINTING_DISH_MEASURED"]:
         assert data_var in pointing_xds.data_vars
-
-
-@pytest.mark.parametrize(
-    "input_target, input_offset, expected_output, expected_error",
-    [
-        (None, None, None, pytest.raises(TypeError, match="NoneType")),
-        (
-            np.array([[[np.pi / 2, np.pi / 2]]]),
-            np.array([[[0.001, 0.0]]]),
-            [[[3.14259, 0.0]]],
-            no_raises(),
-        ),
-        (
-            np.array([[[np.pi / 2, np.pi / 2]]]),
-            np.array([[[0.001, 0.001]]]),
-            [[[3.14259, 1e-3]]],
-            no_raises(),
-        ),
-        (
-            np.array([[[np.pi / 2, np.pi / 2]]]),
-            np.array([[[0.0, 0.0]]]),
-            [[[np.pi, 0.0]]],
-            no_raises(),
-        ),
-        (
-            np.array([[[0.90, 0.35]]]),
-            np.array([[[0.01, 0.001]]]),
-            [[[2.4733, 0.0097366]]],
-            no_raises(),
-        ),
-        (
-            np.array([[[0.90, 0.35]]]),
-            np.array([[[0.0, 0.0]]]),
-            [[[2.4708, 0.0]]],
-            no_raises(),
-        ),
-    ],
-)
-def test__rotate_offset_to_target(
-    input_target, input_offset, expected_output, expected_error
-):
-    with expected_error:
-        result = _rotate_offset_to_target(input_target, input_offset)
-        assert np.allclose(result, expected_output, rtol=1e-5)
